Infer HttpVersion in requests/responses based upon protocol selection (#1321)
Motivation:
HTTP/1.x protocol requires the version is included in the request and
response. When generating requests/response objects in the application
we tend to default to HTTP_1_1 despite the protocol and user
preferences. This may result in additional conversions and confusion
when using version accessors.

Modifications:
- For request/response factories that can derive the real version from
  the transport they should use the real version
- For use cases that don't know which protocol will be used (e.g. Client
  where h1, h2, etc. is requested) use the version most preferred
locally. Worst case additional conversions will be required if the
preferred protocol isn't used.

Result:
HttpVersion in request/response objects will be more accurate and less
conversions will be required.