Require gRPC blocking streaming services to close GrpcPayloadWriter
Motivation:
The HTTP blocking streaming APIs passes ownership of the
HttpPayloadWriter to the user to close. The user may process the stream
on another thread after the handle method has returned, or close the
stream with an exception status. The gRPC blocking streaming handle
method always force closes the stream after the user code returns.

Modifications:
- Remove the force closure in gRPC blocking streaming routing layer
- Fix tests which were not invoking close()

Result:
HTTP and gRPC blocking streaming APIs are more consistent, and the user
has the ability to control when the GrpcPayloadWriter is closed.