Deprecate MultiAddress and Partitioned client builder methods
Motivation:
PartitionedHttpClientBuilder and MultiAddressHttpClientBuilder replicate
methods that exist on SingleAddressHttpClientBuilder as well as
providing a mechanism to directly access the
SingleAddressHttpClientBuilder for each new client to configure it. We
only need one of these mechanisms and the direct access to
SingleAddressHttpClientBuilder reduces duplication and is the more
flexible approach.

Modifications:
- Deprecate all methods on PartitionedHttpClientBuilder and
  MultiAddressHttpClientBuilder which can be otherwise controlled via
  appendClientBuilderFilter(..).
- Deprecate PartitionHttpClientBuilderConfigurator in favor of
  PartitionedHttpClientBuilder.SingleAddressInitializer which is now
  consistent with MultiAddressHttpClientBuilder.SingleAddressInitializer

Result:
Duplicate methods on PartitionedHttpClientBuilder and
MultiAddressHttpClientBuilder have been deprecated indicating the
preferred alternative.

Deprecated methods will be removed in a future release.