Do not add `transfer-encoding: chunked` inside `transform` method
Motivation:

Our current implementation of `transform(TrailersTransformer)` method on
request/response types adds `transfer-encoding: chunked` header with
an assumption that if users use this method, they may add trailers, and the
chunked encoding is required.

However, this method is also used only to access and observe trailers on
the receiver side (e.g., for metrics purposes). As a result, it adds
`transfer-encoding: chunked` header which alters the meaning of the received
message when nothing actually happens.

Modifications:

- Do not add `transfer-encoding: chunked` inside `transform` method, defer this
logic until we prepare a message to be written;
- Improve `addResponseTransferEncodingIfNecessary` and
`addRequestTransferEncodingIfNecessary` to add `transfer-encoding: chunked` if
the message may have trailers;
- `HttpApiConversions#mayHaveTrailers` does not check headers anymore, it's not
necessary;
- Add tests to verify the expected behavior;
- Remove unused `StreamingHttpPayloadHolder#h1TrailersSupported` field;
- Rename `HeaderUtils#shouldAddZeroContentLength` to `responseMayHaveContent`;

Result:

`transform(TrailersTransformer)` does not add `transfer-encoding: chunked` header.