Remove 1ms connect timeout in tests
Motivation:

We added `socketOption(CONNECT_TIMEOUT, 1)` in the past to make some
tests pass on Windows with a comment that Windows uses seconds for the
timeout value. However, Linux uses milliseconds and the `CONNECT_TIMEOUT`
socket option says it have to be in milliseconds. 1ms timeout makes
these tests flaky on Linux and macOS. Let's revisit when we will have
Windows CI pipeline and understand why this socket option is not consistent
on different OS.

Modifications:

- Remove `CONNECT_TIMEOUT` from `H2PriorKnowledgeFeatureParityTest`
and `MultiAddressUrlHttpClientTest`;

Result:

Tests do not fail due to 1ms connect timeout on Linux.