Add Single.zipWith and Single.zip operators
Motivation:
Executing multiple async operations in parallel and combining them in a
type safe way is a common use case, and referred to as zip [1]. We do
not support a zip operator which requires a careful sequence of
operators to replicate.

[1] http://reactivex.io/documentation/operators/zip.html

Modifications:
- Add Single.zipWith
- Add Single.zip overload for 3, 4, and varargs amount of Singles

Result:
Single.zip and Single.zipWith operators exist.