Publisher, Single, Completable onError* operators
Motivation:
ServiceTalk supports a recoverWith operator which allows for error
recovery [1]. However the signature of this method is more complex than
is often required for simple error recovery, and the name implies
"recovery" which is confusing if the return value is intended to
propagate an error.

[1] http://reactivex.io/documentation/operators/catch.html

Modifications:
- Add onError* methods to Publisher, Single, and Completable. The
  signatures are specialized for their respective operation (mapping
  Throwable, swallowing exceptions, recovering with a value) and also
  supports filtering by class type and custom Predicate. This naming
  convention is more commonly used by other JVM based rx implementations
  and is assumed to be more familiar to users.
- Deprecate recoverWith in favor of onErrorResume

Result:
Publisher, Single, and Completable have onError* operators which provide
more APIs tailored toward common error recovery scenarios.