Add Publisher.timeoutTerminal(Duration) operator (#1445)
Motivation:
The existing `Publisher.timeout(Duration)` operator restarts the timer for each received item and will only timeout if no items are emitted during the timeout duration. In some cases the total time for all items to be emitted is of more interest. An operator to timeout based on the entire elapsed time since subscribe is needed.

Modifications:
Add a new Publisher operator, `timeoutTerminal(Duration)`, that terminates the Publisher with a `TimeoutException` if the Publisher does not terminate before the specified time duration.

Result:
A new timeout option is available.