Restore CharSequences.newAsciiString(String) (#1503)
Motivation:
In ServiceTalk 0.39.0 the method
`CharSequences.newAsciiString(CharSequence)` was introduced which
replaced `CharSequences.newAsciiString(String)`. This was a source
compatible, but binary incompatible change. All existing uses of the
`String` overload could be recompiled to use the `CharSequence`
overload. Existing binary code expected the `String` overload to exist
and fails with `NoSuchMethodError` at runtime.
Modifications:
For cases where the libraries and applications using ServiceTalk cannot
be recompiled conveniently the `CharSequences.newAsciiString(String)`
overload is restored.
Result:
Binary compatibility with libraries and applications which use
`CharSequences.newAsciiString(String)` and are compiled against
versions of ServiceTalk before 0.39.0 is restored.