Optimize encoding of HTTP/2 frames to write less on the wire
Motivation:

We can infer from the `PayloadInfo` when HTTP message does not contain
payload body or trailers. In this case, we can write it in a single frame.
Also, we do not need to write `Http2DataFrame`(s) for empty `Buffer`(s)
or process `HttpHeaders` when it's empty.

Modifications:

- Detect if the `HttpMetaData` can be written in a single frame;
- Ignore empty `Buffer`(s) when writing `Http2DataFrame`(s);
- Do not process `HttpHeaders` when they are empty;
- Add tests for new behavior;

Result:

Less `Http2StreamFrame`(s) allocated and written on the wire, results in
+10% RPS, -12% latency for some HTTP/2 use-cases.