Reduce default test timeout on CI to 30s for jUnit4
Motivation:

We set 90s timeout for CI a while ago, when we used a custom CI server.
With GitHub Actions there is no need for that high timeout value. It
only slows down the overall build duration if any timeout happens, but
does not provide much value. The default timeout for jUnit5 tests is
already 30s.

Modifications:

- Change `ServiceTalkTestTimeout.DEFAULT_TIMEOUT_SECONDS` to 30 when
`CI == true`;

Result:

Shorter timeout for jUnit4 tests when build runs on CI.