Improve `NettyConnectionClosing#onClosing()` for HTTP/2
Motivation:

`onClosing()` should be completed as soon as possible to notify LB and
concurrency controller that this connection is going to close soon.
`H2ParentConnectionContext` does not complete `onClosing` until user
closure is invoked on the event-loop (there is no need to wait for that)
and misses a few other pipeline events.

Modifications:

- Complete `onClosing` as soon as we `initiateGracefulClose`;
- Complete `onClosing` for the following pipeline events:
`channelInactive`, `handlerRemoved`, `SslCloseCompletionEvent`,
`ChannelInputShutdownReadComplete`, `ChannelOutputShutdownEvent` because
all of them indicate that the connection is going to close;

Result:

`NettyConnectionClosing#onClosing()` is triggered sooner for HTTP/2
connections.