Allow setting peerHost to null in ClientSslConfig.
Motivation:

There are cases when clients require connecting to a secure server which doesn't
identify itself in the TLS certificate with the same address as used for
connection establishment, or multiple identities can be presented from
diffrent underlying backends. In such case the hostname verification would fail
and would need to be disabled. However, the peerHost which is inferred from the
connection address would unnecessarily cache SSL sessions, which could lead
to issues in session resumption. Therefore, we want to allow explicitly disabling
peerHost inference and either allow setting it to a desired address or to null.

Modifications:

- Added ClientSslConfig#isPeerHostSet method,
- ClientSslConfigBuilder#setPeerHost now accepts null and disables address inference,
- Added test in SslAndNonSslConnectionsTest validating the mentioned scenario,
- Added test in SslAndNonSslConnectionsTest for ensuring hostname validation
  works as expected when hostname is used (existing tests used only IP).

Result:

It is now possible to explicitly use `null` for `peerHost`
and disable its' inference.