Fix ProxyTunnel to always use the address instead of the hostname.
Motivation:

ProxyTunnel was returning the hostname instead of the address, which in some configurations
where the IPv4 is preferred but the DNS resolves in both IPv4 and IPv6 for the loopback hostname,
cause connectivity issues.

Modifications:

Modified the ProxyTunnel to always return the address for the clients to connect to.

Result:

More stable test.