Add SingleThreadedExecutor
Motivation:
For some applications it is important that all signals be handled on
the same thread, sometimes called "Thread Affinity". This is typically
needed because of the use of `ThreadLocal` to pass state between
invocations.

Modifications:
Add a SingleThreadedExecutor implementation that can execute one task
at a time. While concurrent tasks are generally rejected, the
`execute(Runnable)` method provides special behavior if the invoking
thread is the executor worker thread: the `Runnable` will be executed
immediately on the calling thread.

Result:
Another option for handling of signals where thread affinity is
desired.