Defer trailer transformer callbacks on null msg-body to subscription.
Motivation:

When we have a null msg-body and call transform, then the callbacks of the trailer
transformer are invoked on the calling thread rather than the subsciption.

Modifications:

Used defer to defer the invocation.

Result:

Better error handling in case of erroneous trailer transformer (ie. trailers validations)
and deferred execution of the callbacks.