Load balancer should not close existing connections
Motivation:

When dealing with DNS based discovery, the RoundRobinLoadBalancer
eagerly closes connections to hosts not returned in the DNS response. It
makes sense to keep established connections open until the server
decides to close them, but allow traffic shifting by no opening new
connections to unavailable hosts. This change allows configuring the
RoundRobinLoadBalancer to choose the desired behaviour.

Modifications:

* Added a boolean flag to RoundRobinLoadBalancer that allows treating
  unavailable hosts as expired and continue using its' connections.
* Created a Builder for RoundRobinLoadBalancerFactory that allows
  configuring the RoundRobinLoadBalancer with the new flag.

Result:

It is possible to either maintain existing behaviour of
RoundRobinLoadBalancer's flag or to keep connections open to unavailable
hosts (the new default behaviour).