Prevent memory leak of HTTP/1.1 encoding fails
Motivation:

When `HttpObjectEncoder` encodes meta-data or trailers an error may
happen, which will cause a leak of internal `ByteBuf`. It may happen if
any of the encoding object throw or if it tries to expand the size of
that internal `ByteBuf` without success.

Modifications:

- try-catch meta-data and trailers encoding and release the tmp
`ByteBuf` in case of any failures;

Result:

Internal `ByteBuf` does not leak in case of an error in
`HttpObjectEncoder`.