Allow using `null` peer host for TLS and disabling host and port inference and SNI altogether. (#1561)
Allow using `null` peer host for TLS and disabling host and port inference and SNI altogether.

Motivation:

There are cases when clients require connecting to a secure server which doesn't
identify itself in the TLS certificate with the same address as used for
connection establishment, or multiple identities can be presented from
diffrent underlying backends. In such case the hostname verification would fail
and would need to be disabled. However, the peerHost which is inferred from the
connection address would unnecessarily cache SSL sessions, which could lead
to issues in session resumption. Therefore, we want to allow explicitly disabling
peerHost inference and either allow setting it to a desired address or leaving `null`.

Modifications:

- Added `SingleAddressHttpClientBuilder#inferPeerHost` method,
- Added `SingleAddressHttpClientBuilder#inferPeerPort` method,
- Added `SingleAddressHttpClientBuilder#inferSniHostname` method,
- `ClientSslConfigBuilder#setPeerHost` now accepts explicit `null`,
- Added case in `SslAndNonSslConnectionsTest` validating the mentioned scenario,
- Added case in `SslAndNonSslConnectionsTest` for ensuring hostname validation
  works as expected when hostname is used (existing tests used only IP),
- Added case in SniTest validating behaviour for disabled SNI and explicitly
  disabling it in existing `noSni` cases.

Result:

It is now possible to explicitly use `null` for `peerHost` in `ClientSslConfig`
and disable its' inference using `SingleAddressHttpClientBuilder`'s API.