Publisher#flatMapMergeSingle catch an propagate unexpected errors downstream
Motivation:
Publisher#flatMapMergeSingle relies upon the upstream source to catch
unexpected exceptions, cleanup necessary state, and propagate an error
back downstream. However in the event the upstream source is a Single it
will not be able to propagate an error downstream because it has already
sent a terminal signal (multiple terminal signals not allowed by the
RS spec). This means the upstream can at best log/swallow the error and
downstream may not be notified of the failure.

Modifications:
- Publisher#flatMapMergeSingle exception handling is modified to instead
  of propagating the exception upstream, exceptions are caught, upstream
  is cancelled, and the exception is propagated downstream.

Result:
Publisher#flatMapMergeSingle usages where downstream operators may throw
will cancel upstream and propagate and error downstream.