Remove fields from Single (#1622)
`Single` has two fields, `executor` and `shareContextOnSubscribe` that provide
limited value because the contents are unused in all but a few operators. Eliminating
them will reduce allocation and eliminate the held references which could be preventing
GC. This change is also an incremental step in simplifying offloading as `Executor` state
will progressively be reduced to only operators that involve offloading.
Modifications:
Fields are removed from `Single` base and defined locally in sub-classes which have
actual values for the fields.
Result:
Lower allocation, fewer unused local fields.