Disable offloading for `NettyConnectionContext#transportError()` (#1617)
Motivation:

`NettyConnectionContext` is internal API. We do not expect users to use
it directly. If they do, they have to go through the casting of types and take
care about offloading or use it in a non blocking way.
We applied the same strategy for `NettyConnectionContext#onClosing()` in
#1585.

Modifications:

- Update javadoc for `NettyConnectionContext#transportError()`;
- Remove `publishOn(executor)` from its implementation in
`DefaultNettyConnection` and `H2ParentConnectionContext`;

Result:

Internal API of `NettyConnectionContext#transportError()` does not
apply offloading by default.