Fix PublisherFlatMapSingle's demand handling
Motivation:

PublisherFlatMapSingle was recently changed to better handle excpetions, but the changes
affected the way the drain counter was updating, which as a result created wrong demand upstream.

Modifications:

Init the counter inside the loop to reset it everytime there is contention and the lock is re-acquired.

Result:

Proper counting of delivered items and thus proper demand requests.