Enable Pitest plugin for ServiceTalk (#1510)
Motivation:
[PIT](https://pitest.org) provides mutation testing which is a good
measure of the completeness of a project's API tests. As ServiceTalk
has a large and complex API it is worth testing to ensure that the
tests are complete and adequate. The quality of the API and tests can
be improved by testing currently unhandled cases.

Modifications:
A [PIT gradle plugin](https://gradle-pitest-plugin.solidsoft.info)
which provides PIT integration is added to the build. Because PIT
testing is extremely slow the plugin is not currently added to the
quality task.

Result:
The `pitest` task is now available. The reporting by the PITest
plugin is currently limited but is available in the module
directory `build/pitest/index.html` after executing the `pitest`
task.