Removing deprecated methods from MultiAddressHttpClientBuilder
Motivation:

The overridden methods from BaseHttpBuilder have been deprecated and
it's safe to remove them. A `SingleAddressInitializer` serves the
purpose of configuring the individual address' clients.

Modifications:

- Removed inheritance of `HttpClientBuilder` and transitive
  `BaseHttpBuilder` from `MultiAddressHttpClientBuilder`,
- Removed deprecated methods.

Result:

The old methods are gone and the dependency on a base class meant for
single address builders is also removed. The new mechanism using
`SingleAddressInitializer` is the only correct way to configure
individual underlying clients used by `MultiAddressHttpClientBuilder`.