Enable CI for `0.41` branch
Motivation:

`0.41` branch will be used for a while until we stabilize API. We plan
to backport all bux fixes and improvements that do not impact API into
that branch. We need to enable CI in order to provide snapshots of
`0.41.X` versions and validate any PRs that we will need to open for
that branch.

Modifications:
- Add `0.41` to the list of branches for all CI configurations;

Result:

CI will publish snapshots for `0.41` branch and validate all PRs.